+++ ios/Exponent/Versioned/Core/Api/Components/Svg/Elements/RNSVGClipPath.h
@@ -14,4 +14,6 @@
 
 @interface RNSVGClipPath : RNSVGGroup
 
+- (BOOL)isSimpleClipPath;
+
 @end