+++ ios/Exponent/Versioned/Core/Api/Components/Svg/ViewManagers/RNSVGUseManager.m
@@ -19,7 +19,15 @@ RCT_EXPORT_MODULE()
 }
 
 RCT_EXPORT_VIEW_PROPERTY(href, NSString)
-RCT_EXPORT_VIEW_PROPERTY(width, NSString)
-RCT_EXPORT_VIEW_PROPERTY(height, NSString)
+RCT_EXPORT_VIEW_PROPERTY(useheight, RNSVGLength*)
+RCT_EXPORT_VIEW_PROPERTY(usewidth, RNSVGLength*)
+RCT_CUSTOM_VIEW_PROPERTY(height, id, RNSVGUse)
+{
+    view.useheight = [RCTConvert RNSVGLength:json];
+}
+RCT_CUSTOM_VIEW_PROPERTY(width, id, RNSVGUse)
+{
+    view.usewidth = [RCTConvert RNSVGLength:json];
+}
 
 @end